Change decimal values to INT in db_build move ssl into options in db_conn

+++ database/db_build.sql
@@ -6,7 +6,7 @@ CREATE TABLE kitty (
   id SERIAL PRIMARY KEY,
   transaction_date DEFAULT GETDATE(),
   name VARCHAR(200) NOT NULL,
-  transaction_value DECIMAL NOT NULL,
+  transaction_value INT NOT NULL,
 );
 
 INSERT INTO kitty (name, transaction_value)
